Move Handler, APIError to rest-types
