+++ beacon_node/rest_api/src/router.rs
@@ -1,33 +1,22 @@
 use crate::{
-    advanced, beacon,
-    config::{ApiEncodingFormat, Config},
-    consensus,
-    error::ApiError,
-    helpers, lighthouse, metrics, node,
-    response_builder::ResponseBuilder,
-    validator, ApiResult, NetworkChannel,
+    beacon, config::Config, consensus, lighthouse, metrics, node, validator, NetworkChannel,
 };
 use beacon_chain::{BeaconChain, BeaconChainTypes};
 use bus::Bus;
+use environment::TaskExecutor;
 use eth2_config::Eth2Config;
 use eth2_libp2p::{NetworkGlobals, PeerId};
-use hyper::header::{self, HeaderValue};
-use hyper::{body::Bytes, Body, Method, Request, Response, StatusCode};
+use hyper::header::HeaderValue;
+use hyper::{Body, Method, Request, Response};
 use lighthouse_version::version_with_platform;
 use operation_pool::PersistedOperationPool;
 use parking_lot::Mutex;
-use rest_types::Health;
-use serde::Serialize;
+use rest_types::{ApiError, Handler, Health};
 use slog::debug;
-use ssz::Encode;
-use std::future::Future;
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Instant;
-use types::{EthSpec, SignedBeaconBlockHash, Slot};
-
-use environment::TaskExecutor;
-use slog::{info, warn};
+use types::{EthSpec, SignedBeaconBlockHash};
 
 pub struct Context<T: BeaconChainTypes> {
     pub executor: TaskExecutor,
@@ -40,194 +29,6 @@ pub struct Context<T: BeaconChainTypes> {
     pub db_path: PathBuf,
     pub freezer_db_path: PathBuf,
     pub events: Arc<Mutex<Bus<SignedBeaconBlockHash>>>,
-}
-
-struct Handler<T: BeaconChainTypes> {
-    req: Request<()>,
-    body: Body,
-    ctx: Arc<Context<T>>,
-    encoding: ApiEncodingFormat,
-    allow_body: bool,
-}
-
-impl<T: BeaconChainTypes> Handler<T> {
-    pub fn new(req: Request<Body>, ctx: Arc<Context<T>>) -> Result<Self, ApiError> {
-        let (req_parts, body) = req.into_parts();
-        let req = Request::from_parts(req_parts, ());
-
-        let accept_header: String = req
-            .headers()
-            .get(header::ACCEPT)
-            .map_or(Ok(""), |h| h.to_str())
-            .map_err(|e| {
-                ApiError::BadRequest(format!(
-                    "The Accept header contains invalid characters: {:?}",
-                    e
-                ))
-            })
-            .map(String::from)?;
-
-        Ok(Self {
-            req,
-            body,
-            ctx,
-            allow_body: false,
-            encoding: ApiEncodingFormat::from(accept_header.as_str()),
-        })
-    }
-
-    pub fn allow_body(mut self) -> Self {
-        self.allow_body = true;
-        self
-    }
-
-    async fn static_value<V>(self, value: V) -> Result<HandledRequest<V>, ApiError> {
-        // Always check and disallow a body for a static value.
-        let _ = Self::get_body(self.body, false).await?;
-
-        Ok(HandledRequest {
-            value,
-            encoding: self.encoding,
-        })
-    }
-
-    async fn in_core_task<F, V>(self, func: F) -> Result<HandledRequest<V>, ApiError>
-    where
-        V: Send + Sync + 'static,
-        F: Fn(Request<Vec<u8>>, Arc<Context<T>>) -> Result<V, ApiError> + Send + Sync + 'static,
-    {
-        let body = Self::get_body(self.body, self.allow_body).await?;
-        let (req_parts, _) = self.req.into_parts();
-        let req = Request::from_parts(req_parts, body);
-
-        let value = func(req, self.ctx)?;
-
-        Ok(HandledRequest {
-            value,
-            encoding: self.encoding,
-        })
-    }
-
-    async fn in_blocking_task<F, V>(self, func: F) -> Result<HandledRequest<V>, ApiError>
-    where
-        V: Send + Sync + 'static,
-        F: Fn(Request<Vec<u8>>, Arc<Context<T>>) -> Result<V, ApiError> + Send + Sync + 'static,
-    {
-        let ctx = self.ctx;
-        let body = Self::get_body(self.body, self.allow_body).await?;
-        let (req_parts, _) = self.req.into_parts();
-        let req = Request::from_parts(req_parts, body);
-
-        let value = ctx
-            .executor
-            .clone()
-            .handle
-            .spawn_blocking(move || func(req, ctx))
-            .await
-            .map_err(|e| {
-                ApiError::ServerError(format!(
-                    "Failed to get blocking join handle: {}",
-                    e.to_string()
-                ))
-            })??;
-
-        Ok(HandledRequest {
-            value,
-            encoding: self.encoding,
-        })
-    }
-
-    async fn get_body(body: Body, allow_body: bool) -> Result<Vec<u8>, ApiError> {
-        let bytes = hyper::body::to_bytes(body)
-            .await
-            .map_err(|e| ApiError::ServerError(format!("Unable to get request body: {:?}", e)))?;
-
-        if !allow_body && !bytes[..].is_empty() {
-            return Err(ApiError::BadRequest(
-                "The request body must be empty".to_string(),
-            ));
-        } else {
-            Ok(bytes.into_iter().collect())
-        }
-    }
-}
-
-struct HandledRequest<V> {
-    encoding: ApiEncodingFormat,
-    value: V,
-}
-
-impl HandledRequest<String> {
-    pub fn from_string(value: String) -> ApiResult {
-        Self {
-            encoding: ApiEncodingFormat::JSON,
-            value,
-        }
-        .text_encoding()
-    }
-
-    pub fn text_encoding(self) -> ApiResult {
-        Response::builder()
-            .status(StatusCode::OK)
-            .header("content-type", "text/plain; charset=utf-8")
-            .body(Body::from(self.value))
-            .map_err(|e| ApiError::ServerError(format!("Failed to build response: {:?}", e)))
-    }
-}
-
-impl<V: Serialize + Encode> HandledRequest<V> {
-    pub fn all_encodings(self) -> ApiResult {
-        match self.encoding {
-            ApiEncodingFormat::SSZ => Response::builder()
-                .status(StatusCode::OK)
-                .header("content-type", "application/ssz")
-                .body(Body::from(self.value.as_ssz_bytes()))
-                .map_err(|e| ApiError::ServerError(format!("Failed to build response: {:?}", e))),
-            _ => self.serde_encodings(),
-        }
-    }
-}
-
-impl<V: Serialize> HandledRequest<V> {
-    pub fn serde_encodings(self) -> ApiResult {
-        let (body, content_type) = match self.encoding {
-            ApiEncodingFormat::JSON => (
-                Body::from(serde_json::to_string(&self.value).map_err(|e| {
-                    ApiError::ServerError(format!(
-                        "Unable to serialize response body as JSON: {:?}",
-                        e
-                    ))
-                })?),
-                "application/json",
-            ),
-            ApiEncodingFormat::SSZ => {
-                return Err(ApiError::UnsupportedType(
-                    "Response cannot be encoded as SSZ.".into(),
-                ));
-            }
-            ApiEncodingFormat::YAML => (
-                Body::from(serde_yaml::to_string(&self.value).map_err(|e| {
-                    ApiError::ServerError(format!(
-                        "Unable to serialize response body as YAML: {:?}",
-                        e
-                    ))
-                })?),
-                "application/yaml",
-            ),
-        };
-
-        Response::builder()
-            .status(StatusCode::OK)
-            .header("content-type", content_type)
-            .body(body)
-            .map_err(|e| ApiError::ServerError(format!("Failed to build response: {:?}", e)))
-    }
-}
-
-async fn get_body(req: Request<Body>) -> Result<Bytes, ApiError> {
-    hyper::body::to_bytes(req.into_body())
-        .await
-        .map_err(|e| ApiError::ServerError(format!("Unable to get request body: {:?}", e)))
 }
 
 pub async fn route<T: BeaconChainTypes>(
@@ -243,55 +44,35 @@ pub async fn route<T: BeaconChainTypes>(
         let _timer = metrics::start_timer(&metrics::REQUEST_RESPONSE_TIME);
         let ctx = ctx.clone();
         let method = req.method().clone();
-        let handler = Handler::new(req, ctx)?;
+        let executor = ctx.executor.clone();
+        let handler = Handler::new(req, ctx, executor)?;
 
         match (method, path.as_ref()) {
-            /*
-             * Current lighthouse version.
-             */
             (Method::GET, "/node/version") => handler
                 .static_value(version_with_platform())
                 .await?
                 .text_encoding(),
-            /*
-             * The health of the host.
-             */
             (Method::GET, "/node/health") => handler
                 .static_value(Health::observe().map_err(ApiError::ServerError)?)
                 .await?
                 .serde_encodings(),
-            /*
-             * Network syncing status.
-             */
             (Method::GET, "/node/syncing") => handler
                 .allow_body()
                 .in_blocking_task(|_, ctx| node::syncing(ctx))
                 .await?
                 .serde_encodings(),
-            /*
-             * ENR of this node.
-             */
             (Method::GET, "/network/enr") => handler
                 .in_core_task(|_, ctx| Ok(ctx.network_globals.local_enr().to_base64()))
                 .await?
                 .serde_encodings(),
-            /*
-             * Count of peers connected to this node.
-             */
             (Method::GET, "/network/peer_count") => handler
                 .in_core_task(|_, ctx| Ok(ctx.network_globals.connected_peers()))
                 .await?
                 .serde_encodings(),
-            /*
-             * Peer id (libp2p) of this node.
-             */
             (Method::GET, "/network/peer_id") => handler
                 .in_core_task(|_, ctx| Ok(ctx.network_globals.local_peer_id().to_base58()))
                 .await?
                 .serde_encodings(),
-            /*
-             * List of peers connected to this node.
-             */
             (Method::GET, "/network/peers") => handler
                 .in_blocking_task(|_, ctx| {
                     Ok(ctx
@@ -304,59 +85,34 @@ pub async fn route<T: BeaconChainTypes>(
                 })
                 .await?
                 .serde_encodings(),
-            /*
-             * Returns the TCP port number used by libp2p (not discovery).
-             */
             (Method::GET, "/network/listen_port") => handler
                 .in_core_task(|_, ctx| Ok(ctx.network_globals.listen_port_tcp()))
                 .await?
                 .serde_encodings(),
-            /*
-             * Returns a list of multiaddrs (libp2p) for each address that this node is listening
-             * on.
-             */
             (Method::GET, "/network/listen_addresses") => handler
                 .in_blocking_task(|_, ctx| Ok(ctx.network_globals.listen_multiaddrs()))
                 .await?
                 .serde_encodings(),
-            /*
-             * Returns a summary of the head block of the beacon chain.
-             */
             (Method::GET, "/beacon/head") => handler
                 .in_blocking_task(|_, ctx| beacon::get_head(ctx))
                 .await?
                 .all_encodings(),
-            /*
-             * Returns the list of heads of the beacon chain.
-             */
             (Method::GET, "/beacon/heads") => handler
                 .in_blocking_task(|_, ctx| Ok(beacon::get_heads(ctx)))
                 .await?
                 .all_encodings(),
-            /*
-             * Returns a block by slot or root.
-             */
             (Method::GET, "/beacon/block") => handler
                 .in_blocking_task(beacon::get_block)
                 .await?
                 .all_encodings(),
-            /*
-             * Returns the block root in the canonical chain for the given slot.
-             */
             (Method::GET, "/beacon/block_root") => handler
                 .in_blocking_task(beacon::get_block_root)
                 .await?
                 .all_encodings(),
-            /*
-             * Returns the fork of the canonical head.
-             */
             (Method::GET, "/beacon/fork") => handler
                 .in_blocking_task(|_, ctx| Ok(ctx.beacon_chain.head_info()?.fork))
                 .await?
                 .all_encodings(),
-            /*
-             * SSE stream of changes to the canonical head.
-             */
             (Method::GET, "/beacon/fork/stream") => {
                 todo!()
                 /* TODO
@@ -364,25 +120,16 @@ pub async fn route<T: BeaconChainTypes>(
                 beacon::stream_forks::<T>(log, reader)
                 */
             }
-            /*
-             * Returns the genesis time of the canonical head.
-             */
             (Method::GET, "/beacon/genesis_time") => handler
                 .in_blocking_task(|_, ctx| Ok(ctx.beacon_chain.head_info()?.genesis_time))
                 .await?
                 .all_encodings(),
-            /*
-             * Returns the genesis validators root of the canonical head.
-             */
             (Method::GET, "/beacon/genesis_validators_root") => handler
                 .in_blocking_task(|_, ctx| {
                     Ok(ctx.beacon_chain.head_info()?.genesis_validators_root)
                 })
                 .await?
                 .all_encodings(),
-            /*
-             * Return a list of all validators in the canonical head state.
-             */
             (Method::GET, "/beacon/validators") => handler
                 .in_blocking_task(beacon::get_validators)
                 .await?
@@ -477,7 +224,7 @@ pub async fn route<T: BeaconChainTypes>(
                 .await?
                 .serde_encodings(),
             (Method::GET, "/spec/slots_per_epoch") => handler
-                .in_core_task(|_, _| Ok(T::EthSpec::slots_per_epoch()))
+                .static_value(T::EthSpec::slots_per_epoch())
                 .await?
                 .serde_encodings(),
             (Method::GET, "/spec/eth2_config") => handler
@@ -506,7 +253,7 @@ pub async fn route<T: BeaconChainTypes>(
                 .await?
                 .serde_encodings(),
             (Method::GET, "/metrics") => handler
-                .in_blocking_task(metrics::get_prometheus)
+                .in_blocking_task(|_, ctx| metrics::get_prometheus(ctx))
                 .await?
                 .text_encoding(),
             (Method::GET, "/lighthouse/syncing") => handler
