+++ common/rest_types/Cargo.toml
@@ -14,6 +14,13 @@ state_processing = { path = "../../consensus/state_processing" }
 bls = { path = "../../crypto/bls" }
 serde = { version = "1.0.110", features = ["derive"] }
 rayon = "1.3.0"
+hyper = "0.13.5"
+tokio = { version = "0.2.21", features = ["sync"] }
+environment = { path = "../../lighthouse/environment" }
+store = { path = "../../beacon_node/store" }
+beacon_chain = { path = "../../beacon_node/beacon_chain" }
+serde_json = "1.0.52"
+serde_yaml = "0.8.11"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 psutil = "3.1.0"
