+++ common/rest_types/src/lib.rs
@@ -2,20 +2,21 @@
 //!
 //! This is primarily used by the validator client and the beacon node rest API.
 
+mod api_error;
 mod beacon;
 mod consensus;
+mod handler;
 mod node;
 mod validator;
 
+pub use api_error::{ApiError, ApiResult};
 pub use beacon::{
     BlockResponse, CanonicalHeadResponse, Committee, HeadBeaconBlock, StateResponse,
     ValidatorRequest, ValidatorResponse,
 };
-
+pub use consensus::{IndividualVote, IndividualVotesRequest, IndividualVotesResponse};
+pub use handler::{ApiEncodingFormat, Handler};
+pub use node::{Health, SyncingResponse, SyncingStatus};
 pub use validator::{
     ValidatorDutiesRequest, ValidatorDuty, ValidatorDutyBytes, ValidatorSubscription,
 };
-
-pub use consensus::{IndividualVote, IndividualVotesRequest, IndividualVotesResponse};
-
-pub use node::{Health, SyncingResponse, SyncingStatus};
